+++ src/BaseComponent/StringField.tsx
@@ -18,6 +18,7 @@ export default class StringField extends UpFormControl<string> {
             value={this.state.value}
             showError={this.props.showError}
             isRequired={this.props.isRequired}
+            readonly={this.props.isReadOnly && this.props.isReadOnly(this.props.name)}
             onChange={this.handleChangeEventGlobal}
             floatingLabel={this.props.floatingLabel}
           />
@@ -30,6 +31,7 @@ export default class StringField extends UpFormControl<string> {
             showError={this.props.showError}
             isRequired={this.props.isRequired}
             onChange={this.handleChangeEventGlobal}
+            readonly={this.props.isReadOnly && this.props.isReadOnly(this.props.name)}
             floatingLabel={this.props.floatingLabel}
           />
         );
@@ -40,6 +42,7 @@ export default class StringField extends UpFormControl<string> {
             value={this.state.value}
             showError={this.props.showError}
             isRequired={this.props.isRequired}
+            readonly={this.props.isReadOnly && this.props.isReadOnly(this.props.name)}
             onChange={this.handleChangeEventGlobal}
           />
         );
@@ -100,7 +103,7 @@ export default class StringField extends UpFormControl<string> {
             showError={this.props.showError}
             isRequired={this.props.isRequired}
             onChange={this.handleChangeEventGlobal}
-            readonly= {this.props.isReadOnly(this.props.name)}
+            readonly= {this.props.isReadOnly && this.props.isReadOnly(this.props.name)}
             floatingLabel={this.props.floatingLabel}
           />
         );