+++ CHANGELOG.md
@@ -1,3 +1,15 @@
+# React-Schemaform (version: 0.1.79)
+
+## Bug Fixes
+
+- Fix floating label for UpNumber
+
+# React-Schemaform (version: 0.1.78)
+
+## Main changes
+
+- Add new prop isReadonly enabling to handle input state
+
 # React-Schemaform (version: 0.1.77)
 
 ## Bug Fixes